requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pandas
numpy
python-docx
reportlab
matplotlib
openpyxl
pydantic
requests
